+++ renderable/vbText.ts
@@ -1,9 +1,7 @@
 /** https://pixijs.io/pixi-text-style/# */
 import * as PIXI from 'pixi.js';
 import type { LocalizedDictionary, LocalizedTextureMap, TextStyleItem, vbLocalizedObject } from '@vb/core/vbLocalization';
-import { PivotPoint, setPivotRule } from '@vb/core/vbTransform';
 import { c } from '@vb/misc/vbPreset';
-import type { vbGraphicObject } from '@vb/vbGraphicObject';
 import { vbGraphicObjectBase } from '@vb/vbGraphicObject';
 
 
@@ -144,78 +142,5 @@ export type vbTextInitOptions = {
             this.style.align = item.align;
         if (item.break !== undefined)
             this.style.breakWords = item.break;
-    }
-}
-
-
-/**
- * Label is based on an GraphicObject (preferably image or primitive?) with text, \
- * its localization support may change both the graphic object and the text. \
- * NOTE: The names of vbLabel object and its `txt` member variable should be the same
- * in order to properly get the localization.
- */
-export class vbLabel<T extends vbGraphicObject> extends vbGraphicObjectBase(PIXI.Container) implements vbLocalizedObject {
-    bg: T;
-    txt = {} as vbText;
-
-    constructor(bg: T) {
-        super();
-        this.sortableChildren = true;
-        this.bg = bg;
-        this.bg.layer = 0;
-        // this.bg.pivotRule = PivotPoint.Center;
-        this.addChild(bg);
-    }
-
-    get pivotRule() { return this._pivotRule; }
-    set pivotRule(rule: PivotPoint) {
-        this._pivotRule = rule;
-        setPivotRule(this, rule, this.bg.width, this.bg.height);
-    }
-
-    /**
-     * Add a default style text object at the center of graphic object `bg`. \
-     * Set vbLabel object's name by text's name.
-     */
-    addDefaultText(options: vbTextInitOptions) {
-        this.txt = new vbText(options);
-        this.txt.layer = 1;
-        this.txt.pivotRule = PivotPoint.Center;
-        this.txt.position.set(this.bg.width/2, this.bg.height/2);
-        this.addChild(this.txt);
-        // sync name
-        this.name = this.txt.name;
-    }
-
-    addTextObj(obj: vbText) {
-        if (this.txt !== undefined) {
-            this.removeChild(this.txt);
-            this.txt.destroy();
-        }
-        this.txt = obj;
-        this.addChild(this.txt);
-        // sync name
-        this.name = this.txt.name;
-    }
-
-    setText(s: string) {
-        if (this.txt !== undefined) this.txt.text = s;
-    }
-
-    setTextStyle(style: Partial<PIXI.ITextStyle>) {
-        if (this.txt !== undefined) this.txt.style = style;
-    }
-
-    setTextKey(key: string) {
-        this.txt.setKey(key);
-    }
-
-    update(deltaFrame: number) {
-        this.bg.update(deltaFrame);
-        this.txt.update(deltaFrame);
-    }
-
-    localize(dict: LocalizedDictionary, textures: LocalizedTextureMap, item?: TextStyleItem) {
-        this.txt.localize(dict, textures, item);
     }
 }
